pyproject.toml
[project]
name = "the-caverns"
version = "0.1.0"
description = "An exploration ARPG game made in python with pygame."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "moderngl>=5.12.0",
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "pygame-ce==2.5.5",
    "pymunk>=6.11.1",
    "pytmx>=3.32",
]
authors = ["LGgameLAB", "Ben Landon"]

[dependency-groups]
dev = [
    "flake8>=7.1.2",
    "flameprof>=0.4",
    "nuitka>=2.7.11",
    "pep8-naming>=0.14.1",
]

[tool.ruff]
line-length = 100
